ExtremaFindingAlgorithms.h: changed bad comment. FreeImageImage: moved to H3DUtil from H3DAPI. H3DUtil: added initializeH3DUtil and deinitializeH3DUtil functions for initializing of FreeImage when compiling FreeImage static. Needed because Freeimageimage is moved. Also added h3dutil_link_static_externals to use with freeimage. Moved HAVE_ZLIB and HAVE_FREEIMAGE to here instead of H3DAPI.h PixelImage: moved to H3DUtil from H3DAPI. Image.h: added comment about LoadImageFunctions. LoadImageFunctions: moved functions for reading image files from ImageLoaders in H3DAPI to here. RefCountedClass.cpp: added missing default value in constructor. Updated project files to work with this.
git-svn-id: https://www.h3d.org:8090/H3DUtil/trunk@47 875a7764-6f23-0410-b93c-dbbdadb1ea73

+++ H3DUtil/src/Image.cpp
@@ -63,14 +63,14 @@ void Image::getSample( void *value,
   H3DFloat zd = pz - fz;
 
   // interpolate in z
-  H3DUtil::RGBA fff = getPixel( fx, fy, fz );
-  H3DUtil::RGBA ffc = getPixel( fx, fy, cz );
-  H3DUtil::RGBA fcf = getPixel( fx, cy, fz );
-  H3DUtil::RGBA fcc = getPixel( fx, cy, cz );
-  H3DUtil::RGBA cff = getPixel( cx, fy, fz );
-  H3DUtil::RGBA cfc = getPixel( cx, fy, cz );
-  H3DUtil::RGBA ccf = getPixel( cx, cy, fz );
-  H3DUtil::RGBA ccc = getPixel( cx, cy, cz );
+  H3DUtil::RGBA fff = getPixel( (int)fx, (int)fy, (int)fz );
+  H3DUtil::RGBA ffc = getPixel( (int)fx, (int)fy, (int)cz );
+  H3DUtil::RGBA fcf = getPixel( (int)fx, (int)cy, (int)fz );
+  H3DUtil::RGBA fcc = getPixel( (int)fx, (int)cy, (int)cz );
+  H3DUtil::RGBA cff = getPixel( (int)cx, (int)fy, (int)fz );
+  H3DUtil::RGBA cfc = getPixel( (int)cx, (int)fy, (int)cz );
+  H3DUtil::RGBA ccf = getPixel( (int)cx, (int)cy, (int)fz );
+  H3DUtil::RGBA ccc = getPixel( (int)cx, (int)cy, (int)cz );
   
   H3DUtil::RGBA i1 = fff * (1-zd) + ffc * zd;
   H3DUtil::RGBA i2 = fcf * (1-zd) + fcc * zd;
